Initial implementation
Initial implementation of the challenge. It doesn't contain any tests because of time constraints. Resolving HTTP redirects took quite a bit of effort that could have gone into the tests if the requirements weren't that it works with two websites that both use redirects.

Some further optimization might be advised, but the application works as intended.